CONCPP-23 CONCPP-22 CONCPP-12
CONCPP-23 and CONCPP-12 are connected, since good part of warnings were
caused by the SQLString class, and its redesign eleminated them among other
things. CONCPP-22 changes could probably be separated from others, but changed files
could also contain changes for other tickets. And still can be fairly
easy separated in case of need.
SQLString now hides internal string implementation and does not provide
the operator for transparent cast to std::string. While const char*
stays in place. Depending on the compiler, a construction
std::string var= <method_call_returning_SQLString>;
may work or may work not.
std::string var(<method_call_returning_SQLString>);
seems to be more portable.
But in the particular case of a blob fetched with
the getString, using the SQLString variable to accept the value is probably the only
choice(because of possible \0 bytes in the string)

Move constructor has been added.

+++ src/ColumnDefinition.h
@@ -47,7 +47,7 @@ public:
   virtual SQLString getOriginalName() const=0;
   virtual short getCharsetNumber() const=0;
   virtual SQLString getCollation() const=0;
-  virtual int64_t getLength() const=0;
+  virtual uint32_t getLength() const=0;
   virtual int64_t getPrecision() const=0;
   virtual int32_t getDisplaySize() const=0;
   virtual uint8_t getDecimals() const=0;
