+++ src/parameters/StringParameter.h
@@ -46,7 +46,7 @@ public:
   bool isNullData() const;
   bool isLongData();
   void* getValuePtr() { return const_cast<void*>(static_cast<const void*>(stringValue.c_str())); }
-  unsigned long getValueBinLen() const { return stringValue.length(); }
+  unsigned long getValueBinLen() const { return static_cast<unsigned long>(stringValue.length()); }
   };
 }
 }
